wip - Migrate basic constraint, collection constraints
