+++ packages/overture-schema-core/src/overture/schema/core/types.py
@@ -3,7 +3,7 @@ from typing import Annotated, NewType
 
 from pydantic import Field
 
-from overture.schema.foundation.primitive import float64, int32, pct
+from overture.schema.foundation.primitive import int32, pct
 from overture.schema.validation.constraints import (
     CountryCodeConstraint,
     LinearReferenceRangeConstraint,
