+++ packages/overture-schema-foundation/src/overture/schema/foundation/constraint/__init__.py
@@ -1,0 +1,28 @@
+"""
+constraint
+==========
+Reusable constraints.
+
+todo - vic
+
+
+Modules
+-------
+constraint : module
+    Base constraint type.
+collection : module
+    Constraints for collections
+
+"""
+
+from .collection import (
+    CollectionConstraint,
+    UniqueItemsConstraint,
+)
+from .constraint import Constraint
+
+__all__ = [
+    "Constraint",
+    "CollectionConstraint",
+    "UniqueItemsConstraint",
+]
