+++ packages/overture-schema-places-theme/src/overture/schema/places/place/models.py
@@ -17,7 +17,7 @@ from overture.schema.core.types import (
     PhoneNumber,
     WikidataId,
 )
-from overture.schema.core.validation import (
+from overture.schema.foundation.constraint import (
     UniqueItemsConstraint,
 )
 from overture.schema.foundation.primitive import (
