+++ packages/overture-schema-validation/src/overture/schema/validation/__init__.py
@@ -27,6 +27,8 @@ Usage:
         tags: Annotated[list[str], UniqueItemsConstraint()]
 """
 
+from overture.schema.foundation.constraint import UniqueItemsConstraint
+
 from .constraints import (
     CategoryPatternConstraint,
     ConfidenceScoreConstraint,
@@ -40,7 +42,6 @@ from .constraints import (
     PhoneNumberConstraint,
     RegionCodeConstraint,
     StringConstraint,
-    UniqueItemsConstraint,
     WhitespaceConstraint,
     WikidataConstraint,
 )
